Move auxflash core implementation into humility-auxflash
