+++ cmd/auxflash/Cargo.toml
@@ -12,12 +12,10 @@ tlvc.workspace = true
 anyhow.workspace = true
 clap.workspace = true
 colored.workspace = true
-indicatif.workspace = true
 log.workspace = true
 parse_int.workspace = true
 
 humility.workspace = true
+humility-auxflash.workspace = true
 humility-cmd.workspace = true
 humility-cli.workspace = true
-humility-hiffy.workspace = true
-humility-idol.workspace = true
