+++ cmd/flash/src/lib.rs
@@ -35,6 +35,7 @@
 use anyhow::{Context, Result, anyhow, bail};
 use clap::{CommandFactory, Parser};
 use humility::{core::Core, hubris::*};
+use humility_auxflash::AuxFlashHandler;
 use humility_cli::{
     Cli, {ExecutionContext, Subcommand},
 };
@@ -315,15 +316,14 @@ fn get_image_state(
 
         // Note that we only run this check if we pass the image ID check;
         // otherwise, the Idol / hiffy memory maps are unknown.
-        let mut worker =
-            match cmd_auxflash::AuxFlashHandler::new(hubris, core, 15_000) {
-                Ok(w) => w,
-                Err(e) => {
-                    // Halt the core before returning!
-                    core.halt()?;
-                    return Err(e);
-                }
-            };
+        let mut worker = match AuxFlashHandler::new(hubris, core, 15_000) {
+            Ok(w) => w,
+            Err(e) => {
+                // Halt the core before returning!
+                core.halt()?;
+                return Err(e);
+            }
+        };
         let r = match worker.active_slot() {
             Ok(Some(s)) => {
                 humility::msg!("verified auxflash in slot {s}");
@@ -462,7 +462,7 @@ fn program_auxflash(
     core: &mut dyn Core,
     data: &[u8],
 ) -> Result<()> {
-    let mut worker = cmd_auxflash::AuxFlashHandler::new(hubris, core, 15_000)?;
+    let mut worker = AuxFlashHandler::new(hubris, core, 15_000)?;
 
     // At this point, we've already rebooted into the new image.
     //
